requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
gunicorn
line-bot-sdk
openai
PyYAML
